Fix read in XmppOverTls test
The code assumed that the read() method would always either fill the
buffer or otherwise there is an error. While it is in practice very
likely that this is the case, it is fragile to rely on that behavior.

Instead the code should handle the case where read() may only return a
few chars, but not enough to fill the buffer.